+++ chatbot_connectors/__init__.py
@@ -29,7 +29,7 @@ from chatbot_connectors.exceptions import (
 )
 from chatbot_connectors.factory import ChatbotFactory
 
-__version__ = "0.1.0"
+__version__ = "0.9.0"
 
 __all__ = [
     "Chatbot",