+++ chatbot_connectors/implementations/__init__.py
@@ -3,6 +3,7 @@
 from .botslovers import BotsloversChatbot
 from .comunidad_madrid import ComunidadMadridChatbot
 from .custom import CustomChatbot
+from .langgraph import LangGraphChatbot
 from .metro_madrid import MetroMadridChatbot
 from .millionbot import MillionBot
 from .rasa import RasaChatbot
@@ -13,6 +14,7 @@ __all__ = [
     "ChatbotTaskyto",
     "ComunidadMadridChatbot",
     "CustomChatbot",
+    "LangGraphChatbot",
     "MetroMadridChatbot",
     "MillionBot",
     "RasaChatbot",